+++ src/Pos.Infrastructure/Migrations/20260816205847_PermiteOrdenesPendientesMercadoPago.cs
@@ -1,0 +1,66 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace Pos.Infrastructure.Migrations
+{
+    /// <inheritdoc />
+    public partial class PermiteOrdenesPendientesMercadoPago : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AlterColumn<string>(
+                name: "ProviderPaymentId",
+                schema: "pos",
+                table: "mercado_pago_order",
+                type: "character varying(80)",
+                maxLength: 80,
+                nullable: true,
+                oldClrType: typeof(string),
+                oldType: "character varying(80)",
+                oldMaxLength: 80);
+
+            migrationBuilder.AlterColumn<string>(
+                name: "ProviderOrderId",
+                schema: "pos",
+                table: "mercado_pago_order",
+                type: "character varying(80)",
+                maxLength: 80,
+                nullable: true,
+                oldClrType: typeof(string),
+                oldType: "character varying(80)",
+                oldMaxLength: 80);
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AlterColumn<string>(
+                name: "ProviderPaymentId",
+                schema: "pos",
+                table: "mercado_pago_order",
+                type: "character varying(80)",
+                maxLength: 80,
+                nullable: false,
+                defaultValue: "",
+                oldClrType: typeof(string),
+                oldType: "character varying(80)",
+                oldMaxLength: 80,
+                oldNullable: true);
+
+            migrationBuilder.AlterColumn<string>(
+                name: "ProviderOrderId",
+                schema: "pos",
+                table: "mercado_pago_order",
+                type: "character varying(80)",
+                maxLength: 80,
+                nullable: false,
+                defaultValue: "",
+                oldClrType: typeof(string),
+                oldType: "character varying(80)",
+                oldMaxLength: 80,
+                oldNullable: true);
+        }
+    }
+}