+++ src/Pos.Integrations.MercadoPago/Class1.cs
@@ -1,6 +1,0 @@
-﻿namespace Pos.Integrations.MercadoPago;
-
-public class Class1
-{
-
-}